Phase 3: Refactor VM dispatch to use protocol-based dispatch
- RuntimeOps: IsFalsy → IVMTruthiness, IsEqual → IVMEquatable,
  Stringify → IVMStringifiable, Add/Sub/Mul/Div/Negate → IVMArithmetic
  (forward+reverse dispatch), Compare → IVMComparable
- TypeOps: ExecuteTypeOf → IVMTyped, GetFieldValue → IVMFieldAccessible,
  SetFieldValue → IVMFieldMutable
- ControlFlow: ExecuteIterPrep/ExecuteIterLoop → IVMIterable/IVMIterator
- Fix Duration.totalMs returning double instead of long
- NaN equality already IEEE 754 compliant (no change needed)
- All 6294 tests pass (2 pre-existing flaky network failures)

+++ Stash.Tests/Bytecode/TypeProtocolTests.cs
@@ -60,7 +60,7 @@ public class TypeProtocolTests
         var dur = new StashDuration(5000); // 5 seconds
         bool found = ((IVMFieldAccessible)dur).VMTryGetField("totalMs", out StashValue value, null);
         Assert.True(found);
-        Assert.Equal(5000.0, value.AsFloat);
+        Assert.Equal(5000L, value.AsInt);
     }
 
     [Fact]
